+++ features/auth/src/main/kotlin/com/boswelja/truemanager/auth/DI.kt
@@ -2,7 +2,7 @@ package com.boswelja.truemanager.auth
 
 import com.boswelja.truemanager.auth.serverstore.AuthenticatedServersStore
 import com.boswelja.truemanager.auth.serverstore.AuthenticatedServersStoreImpl
-import com.boswelja.truemanager.auth.ui.AuthViewModel
+import com.boswelja.truemanager.auth.ui.addserver.AddServerViewModel
 import org.koin.androidx.viewmodel.dsl.viewModelOf
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.bind
@@ -11,5 +11,5 @@ import org.koin.dsl.module
 val authModule = module {
     singleOf(::AuthenticatedServersStoreImpl) bind AuthenticatedServersStore::class
 
-    viewModelOf(::AuthViewModel)
+    viewModelOf(::AddServerViewModel)
 }
